LIBFCREPO-760. Moved STOMP-specific code into a plastron.stomp module.
https://issues.umd.edu/browse/LIBFCREPO-760
